Separate sync schema and db models
